+++ src/Locales.php
@@ -1,0 +1,56 @@
+<?php
+
+/**
+ * Title: Mollie locales
+ * Description:
+ * Copyright: Copyright (c) 2005 - 2015
+ * Company: Pronamic
+ * @author Remco Tolsma
+ * @version 1.1.0
+ */
+class Pronamic_WP_Pay_Mollie_Locales {
+	/**
+	 * German (Deutsch)
+	 *
+	 * @var string
+	 * @see https://www.mollie.com/nl/docs/payments
+	 * @see http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
+	 */
+	const DE = 'de';
+
+	/**
+	 * English
+	 *
+	 * @var string
+	 * @see https://www.mollie.com/nl/docs/payments
+	 * @see http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
+	 */
+	const EN = 'en';
+
+	/**
+	 * French
+	 *
+	 * @var string
+	 * @see https://www.mollie.com/nl/docs/payments
+	 * @see http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
+	 */
+	const FR = 'fr';
+
+	/**
+	 * Spanish (Español)
+	 *
+	 * @var string
+	 * @see https://www.mollie.com/nl/docs/payments
+	 * @see http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
+	 */
+	const ES = 'es';
+
+	/**
+	 * Dutch (Nederlands)
+	 *
+	 * @var string
+	 * @see https://www.mollie.com/nl/docs/payments
+	 * @see http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
+	 */
+	const NL = 'nl';
+}