+++ src/PaymentRequest.php
@@ -3,7 +3,7 @@
 /**
  * Title: Mollie payment request
  * Description:
- * Copyright: Copyright (c) 2005 - 2014
+ * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
  * @author Remco Tolsma
  * @version 1.0.0
@@ -36,7 +36,7 @@ class Pronamic_WP_Pay_Gateways_Mollie_PaymentRequest {
 			'method'      => $this->method,
 			'redirectUrl' => $this->redirect_url,
 			'metadata'    => $this->meta_data,
-			'locale'      => strtolower( $this->locale ),
+			'locale'      => $this->locale,
 		);
 
 		// Array filter will remove values NULL, FALSE and empty strings ('')