+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 1.1.0
+*	Improved support for unsupported Mollie locales.
+*	Added support for the SOFORT Banking payment method.
+
 ## 1.0.0
 *	First release.